// src/app/services/api/atm/model/administrationDoctorDto.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface AdministrationDoctorDto {
  actualDateTime?: Date;
  administrationNote?: string;
  administrationRoute?: string;
  dateTime?: Date;
  drugName?: string;
  measurementUnit?: string;
  order?: number;
  ptComponentNote?: string;
  quantity?: number;
  readingNoteStatus?: AdministrationDoctorDto.ReadingNoteStatusEnum;
  status?: AdministrationDoctorDto.StatusEnum;
  uuid?: string;
}
export namespace AdministrationDoctorDto {
  export type ReadingNoteStatusEnum = 'TO_READ' | 'READ' | 'NOT_AVAILABLE';
  export const ReadingNoteStatusEnum = {
    TOREAD: 'TO_READ' as ReadingNoteStatusEnum,
    READ: 'READ' as ReadingNoteStatusEnum,
    NOTAVAILABLE: 'NOT_AVAILABLE' as ReadingNoteStatusEnum
  };
  export type StatusEnum = 'TO_PREPARE' | 'TO_ADMINISTER' | 'NOT_ADMINISTERED_MISSED' | 'NOT_ADMINISTERED_REJECTED' | 'NOT_ADMINISTERED_TIME_WINDOW_EXPIRED' | 'UNKNOWN' | 'ADMINISTERED';
  export const StatusEnum = {
    TOPREPARE: 'TO_PREPARE' as StatusEnum,
    TOADMINISTER: 'TO_ADMINISTER' as StatusEnum,
    NOTADMINISTEREDMISSED: 'NOT_ADMINISTERED_MISSED' as StatusEnum,
    NOTADMINISTEREDREJECTED: 'NOT_ADMINISTERED_REJECTED' as StatusEnum,
    NOTADMINISTEREDTIMEWINDOWEXPIRED: 'NOT_ADMINISTERED_TIME_WINDOW_EXPIRED' as StatusEnum,
    UNKNOWN: 'UNKNOWN' as StatusEnum,
    ADMINISTERED: 'ADMINISTERED' as StatusEnum
  };
}