// src/app/services/api/atm/model/administrationsByDoctorDto.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AdministrationDoctorDto } from './administrationDoctorDto';
import { LocalTime } from './localTime';
import { PatientDto } from './patientDto';


export interface AdministrationsByDoctorDto { 
    administrations?: Array<AdministrationDoctorDto>;
    dateRequest?: string;
    dateTimeRequest?: Date;
    patientDto?: PatientDto;
    timeRequest?: LocalTime;
}